requirements.txt
django==3.1.1
djangorestframework>=3.11.1,<3.12.0
django-filter==2.3.0
djangorestframework-filters==0.11.1
djangorestframework-jwt>=1.11.0,<1.12.0
django-jazzmin>=2.2.11,<2.3.0
django-import-export>=2.4.0,<2.5.0
django-resized>=0.3.11,<0.4.0
django-cors-headers>=3.5.0,<3.6.0

psycopg2>=2.8.6,<2.9.0
Pillow
channels>=2.4.0,<2.5.0
channels-redis>=3.1.0,<3.2.0
drf-yasg>=1.17.1,<1.18.0

flake8>=3.6.0,<3.7.0
